add artifacts customize value, update artifacts.xml, profits.xml
git-svn-id: https://fheroes2.svn.sourceforge.net/svnroot/fheroes2/trunk@2201 4f904287-c015-0410-838d-95d3a2656228

+++ fheroes2/src/fheroes2/resource/artifact.h
@@ -180,7 +180,7 @@ public:
     static artifact_t Rand(level_t);
     static u8 GetLevel(artifact_t);
     static u8 GetLevelLoyalty(artifact_t);
-    static u8 GetExtraValue(artifact_t);
+    static u16 GetExtraValue(artifact_t);
 
     static const char* GetName(artifact_t);
     static const char* GetDescription(artifact_t);
